backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-migrate
python-dotenv
google-generativeai
pyjwt
flask-bcrypt
